requirements.txt
Django==4.2.7
djongo==1.3.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==3.12.3
sqlparse==0.2.4
pytz==2023.3
dnspython==2.4.2